restart plan extended

+++ src/main/java/de/grimmpp/AppManager/service/ServicePlanAppRestarter.java
@@ -1,11 +1,9 @@
 package de.grimmpp.AppManager.service;
 
 import de.grimmpp.AppManager.model.cfClient.Application;
+import de.grimmpp.AppManager.model.cfClient.ApplicationInstances;
 import de.grimmpp.AppManager.model.cfClient.Resource;
-import de.grimmpp.AppManager.model.database.Binding;
-import de.grimmpp.AppManager.model.database.BindingRepository;
-import de.grimmpp.AppManager.model.database.ServiceInstance;
-import de.grimmpp.AppManager.model.database.ServiceInstanceRepository;
+import de.grimmpp.AppManager.model.database.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,6 +23,9 @@ public class ServicePlanAppRestarter implements IServicePlan {
     @Autowired
     private BindingRepository bindingRepo;
 
+    @Autowired
+    private ParameterRepository parameterRepo;
+
     @Override
     public void run() throws IOException {
 
@@ -33,12 +34,20 @@ public class ServicePlanAppRestarter implements IServicePlan {
                 String appUrl = cfClient.buildUrl(CfClient.URI_SINGLE_APP, b.getApplicationId());
                 Resource<Application> app = cfClient.getResource(appUrl, Application.class);
 
-                //TODO: time check is missing
-
                 if (app.getEntity().getState().equals("STARTED")) {
-                    for(int i=0; i<app.getEntity().getInstances();i++) {
-                        String instanceUrl = cfClient.buildUrl(CfClient.URI_APP_INSTANCE, b.getApplicationId(), String.valueOf(i));
-                        cfClient.deleteResource(instanceUrl);
+                    Parameter p = parameterRepo.findByReferenceAndKey(b.getServiceInstanceId(), "time");
+                    long time = 0; //TODO: to be set with value from parameter
+
+                    String aiUrl = cfClient.buildUrl(CfClient.URI_APP_INSTANCES, b.getApplicationId());
+                    ApplicationInstances ais = cfClient.getObject(aiUrl, ApplicationInstances.class);
+
+                    boolean isAppExpired = ais.values().stream().anyMatch(i -> i.getUptime().equals(time));
+
+                    if (isAppExpired) {
+                        for (int i = 0; i < app.getEntity().getInstances(); i++) {
+                            String instanceUrl = cfClient.buildUrl(CfClient.URI_APP_INSTANCE, b.getApplicationId(), String.valueOf(i));
+                            cfClient.deleteResource(instanceUrl);
+                        }
                     }
                 }
             }
